Fix tests.
go version go1.7.3 puts the test executable in a temporary directory such as
/tmp/go-build392344006/github.com/modocache/gover/gover/_test, which does not
contain the _fixtures directory, and as a result most of the tests fail.

However it does run the test with CWD set to the source directory, so it's
enough to just look at _fixtures in the current directory.